Use RAII in PageView (#4792)
* Use unique_ptr for PageView::eraser
* Use unique_ptr for PageView::inputHandler

+++ src/core/gui/PageViewFindObjectHelper.h
@@ -39,7 +39,8 @@ public:
 
         if (multiLayer) {
             size_t initialLayer = this->view->getPage()->getSelectedLayerId();
-            for (int layerNo = this->view->getPage()->getLayers()->size() - 1; layerNo >= 0; layerNo--) {
+            for (int layerNo = static_cast<int>(this->view->getPage()->getLayers()->size() - 1); layerNo >= 0;
+                 layerNo--) {
                 Layer* layer = this->view->getPage()->getLayers()->at(layerNo);
                 this->view->getXournal()->getControl()->getLayerController()->switchToLay(layerNo + 1);
                 if (checkLayer(layer)) {
